Sort components (so that "main" is always first); fixes reprepro's default guessed component.
reprepo's default guessed component is the first configured for
the distribution; django order, per default, would put contrib
first, so packages would be build for contrib instead of main.

This fixes this, and also gives the more used-to order "main
contrib non-free" for all component listings.

This also adds a check for the (mandatory) component "main".