api(migrate,remove,port): Add (optional) arg '--version', and generally handle multiple per-dist package versions (via components).
reprepro does only hold one version of a package per
distribution; the only exception is when they are installed in
different components.

This might happen when packages are manually installed via
reprepro (with '-C'), or when a package in it's life cycle at
some point changes its 'Section' in debian/control.

Though it's usually a rare case, we should be able to deal with
that anyway.

First, we add "--version" to API calls migrate, remove and port
to be able to select a specific package version if need be.

Second, the search results for show is fixed to deal with
multi-versions; each package in differnent components get its
own entry.

Third, all the web API calls for port, remove and migrate are
updated to always include "version", so these package can be
dealt with individiually.