Improve status model handling (Fixes long waits on start-up, some possible status glitches).
We give status 'prepared' some more 'substatuses' based on
'check' or changes to the model now: check timestamp, no check,
changed, failed, failed_reactivate.

Checks are usually not run once checked. On changes through
django admin, the check is invalidated as "changed", and
"prepare" takes care of syncing the changed data in the model
with the associated data.

Also, some glitches are removed that could happen when you
change chroots or repositories that are already prepared, which
might have led to stray associated data.

You can force model check via the "check" action in the django
admin; also, on mini-buildd start (not daemon start), check on
the Daemon model with all dependencies is forced.

check() on models has been improved for all status models; most
importantly. chroot models will now check that we can actually
create a snapshot, and repository models with just do a reprepro
check. The latter used to be just a re-dump of configuration and
repepro export for simplicity, but this could take really a long
time (especially if you have configured quite a number of dists
with rollback support).