chroots: On debootstrap, generate and use keyring from associated source to authenticate archive (General fix; especially fixes etch chroot prepare under sid).
Keyring "keyring.gpg" is generated in the chroot dir, and then
we use the '--keyring' option to specify our custom keyring.

Previously, debootstrap would just check per default against the
keyring in installed debian-archive-keyring, which (usually
works [*] but) is not the right thing for mini-buildd as we
already have keys associated with the source.

[*] Currently fixes, f.e., 'etch' chroot prepare under current
'sid', where the etch release keys are deprecated in
debian-archive-keyring.