django 1.5: Set ALLOWED_HOSTS to '*' (omg). (Fixes any http requests when not started with --debug=django).
With django 1.5, the 'ALLOWED_HOSTS' whitelist setting is
mandatory.

Up to now, I only tested locally on sid with
'--debug=django' (blush), where this does not matter.

As we do not have anything useful to set here, we allow from any
host for the time being, bringing back the old behavior.