Sources: Move the archive check code from prepare to check (Fixes: Won't find new archives without remove).
The Source Model is a mock status model, i.e., it has no
external prepared data. So we just use prepare() to check the
correctness of the data without archives (i.e., has >0 apt
keys), and check() to (newly) associate the source with at least
one archive.

Thusly, the scheduled or forced checks on the source will also
update the "fastest" archive, or select a newly added working
one.