builder+packager: Install dsc directly, don't include source in arch-all builds any more.
We relied on the archall changes to get the dsc installed; now
we just install the dsc directly, and the archall-builder does
include the dsc/orig in the changes any more.

This fixes build failures in case there is nothing to build on
the 'archall' arch, which then comes back with 'skipped' and no
changes at all.

It also avoids moving the dsc and orig tarball back from the
builder in the "archall" build result.