Wizard(Distribution): Really always avoid creation of "duplicate" instances.
Since the workaround for dists not having the lintian
'--suppress' option, chroots for lenny and etch are created with
'lintian=runonly' option by the wizard (before, just default
values).

However, using get_or_create() with that extra option might
produce another Distribution with the same codename.

get_or_create() must not be used thusly, only the codename must
be used as arg to always avoid this.